Error handling, change tool cost type, and update editor script

+++ Assets/Scripts/Minigames/ShipRepairs/RepairsConstants.cs
@@ -17,6 +17,6 @@ public static class RepairsConstants
     public static string FailureMessage = "Failure!";
 
     // Tools
-    public static long CostPerTool = 100L;
+    public static int CostPerTool = 100;
 
 }
